api/requirements.txt

fastapi
uvicorn
requests
gitpython
python-dotenv
pydantic
zhipuai
